requirements.txt
Flask
flask_wtf
mysql-connector-python 
bcrypt
Flask-Login
email_validator